30_05_23 Version 1.0
Version 1.0 includes the first work package done these last weeks: downlodad of the data files and the preprocessgin and modeling of what it's supposed to be a minimun valuable project.
Includes all new files: demand, generation, weather...
NOtebooks for solar and eolic preprocessing and modelling.
Notebook with formulas that have been created.

It is expected to migrate all this information to Kedro in future updates